Updated and renamed an existing test
